Update README.md
The purpose of this analysis is to allow participants of the research done by Improbable Beef, a food startup attempting to create the newest variation of a plant-based burger including other microbiological species. One form of bacteria they potentially believe can resemble the taste of a traditional burger is found in the human belly button. Therefore, when researchers have found a useable bacteria, participants can check out the web app created to find their ID and see if the bacteria found in their belly button is a match based on the given bubble/bar chart as well as a gauge chart showing their wash rate.